requirements.txt
appdirs==1.4.3
asgiref==3.2.3
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
dj-rest-auth==1.0.9
Django==3.0.8
django-allauth==0.42.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-pwa==1.0.10
django-suit==0.2.28
django-tag-parser==3.2
djangorestframework==3.11.0
docutils==0.16
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jsmin==2.2.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.2
mccabe==0.6.1
numpy==1.19.0
oauthlib==3.1.0
opencv-python==4.3.0.36
pandas==1.0.5
pep8==1.7.1
Pillow==7.2.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2018.5
requests==2.23.0
requests-oauthlib==1.3.0
scipy==1.5.1
seaborn==0.10.1
six==1.14.0
sqlparse==0.2.4
toml==0.10.1
urllib3==1.25.8
virtualenv==20.0.5
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
whitenoise==5.1.0
wrapt==1.11.2
